Fix PR issues
