--- ImageFeed/Constants/DateConstants.swift
@@ -1,4 +1,0 @@
-enum DateConstants {
-    static let defaultDateFormat = "dd MMMM yyyy"
-    static let defaultLocale = "ru_RU"
-}
