--- ImageFeed/Constants/ErrorConstants.swift
@@ -1,4 +1,0 @@
-enum ErrorConstants {
-    static let failedCast = "Failed to cast ImagesListCell"
-    static let failedImage = "Failed to load image"
-}
