--- ImageFeed/Constants/ImageConstants.swift
@@ -1,4 +1,0 @@
-enum ImageConstants {
-    static let activeLike = "Active Like"
-    static let noActiveLike = "No Active Like"
-}
