--- ImageFeed/Constants/UIConstants.swift
@@ -1,6 +1,0 @@
-import UIKit
-
-enum UIConstants {
-    static let tableViewContentInsets = UIEdgeInsets(top: 12, left: 0, bottom: 12, right: 0)
-    static let cellImageInsets = UIEdgeInsets(top: 4, left: 16, bottom: 4, right: 16)
-}
